+++ src/main/java/org/folio/des/config/kafka/KafkaConfiguration.java
@@ -21,9 +21,7 @@ import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.core.ProducerFactory;
 import org.springframework.kafka.support.serializer.JsonDeserializer;
 import org.springframework.kafka.support.serializer.JsonSerializer;
-import org.springframework.stereotype.Component;
 
-@Component
 @Configuration
 @RequiredArgsConstructor
 public class KafkaConfiguration {