+++ src/main/java/org/folio/des/service/config/impl/BursarFeesFinesExportConfigService.java
@@ -18,8 +18,6 @@ import org.folio.des.service.config.ExportConfigDomainEventService;
 import org.folio.des.validator.ExportConfigValidatorResolver;
 import org.springframework.data.domain.PageRequest;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import lombok.extern.log4j.Log4j2;
 
 @Log4j2
@@ -29,9 +27,9 @@ public class BursarFeesFinesExportConfigService extends BaseExportConfigService 
 
   public BursarFeesFinesExportConfigService(ExportConfigRepository repository, DefaultExportConfigMapper defaultExportConfigMapper,
                                             ExportConfigMapperResolver exportConfigMapperResolver, ExportConfigValidatorResolver exportConfigValidatorResolver,
-                                            ExportConfigDomainEventService exportConfigDomainEventService, ObjectMapper objectMapper,
+                                            ExportConfigDomainEventService exportConfigDomainEventService,
                                             BursarExportScheduler bursarExportScheduler) {
-    super(repository, defaultExportConfigMapper, exportConfigMapperResolver, exportConfigValidatorResolver, exportConfigDomainEventService, objectMapper);
+    super(repository, defaultExportConfigMapper, exportConfigMapperResolver, exportConfigValidatorResolver, exportConfigDomainEventService);
     this.bursarExportScheduler = bursarExportScheduler;
   }